Use a better way to simulate the gas sensor
